Add CreatedByUser in each of the model and modify the controllers
Add CreatedByUser field in all model so that only entries created by the user will be displayed. Modify the controllers in the web app to add a Where clause when reading entries from the database. During creating a new record, manually pass the logged in email address.

Modify the web API controllers by adding an extra parameter for the email address to filter the returned results.