Refine the speech recognition logic and move files to root folder
I have moved the script files to root folder so I can debug the code easily. Also, I have moved the image resources there because I noticed it is not being recognized by the web app if I put in a separate Resources folder. I am still finalizing the speech to text part, I am thinking to have a single microphone button for all the text fields to simplify the logic instead of having to switch the mic on and off for different subsection texts.